fix typos and mnove classes folder
